+++ CHANGELOG.md
@@ -7,9 +7,10 @@
 
 ### Changed
 - version to 1.1.3-SNAPSHOT
+- Scala 2.12, Spark 2.4, JTS from locationtech v1.16
 
 ### Fixed
-- ???
+- tuple structure bug in NearestD case
 
 ### Removed
 - ???