+++ CHANGELOG.md
@@ -1,13 +1,30 @@
 # Changelog
 
 ## [Unreleased]
+
+### Added
+- ???
+
+### Changed
+- ???
+
+### Fixed
+- ???
+
+### Removed
+- ???
+
+### Deprecated
+- ???
+
+## [1.2.1] - 2019-04-17
 Fixed some bugs, added tests, build Spark 2.4 + Scala 2.12 version.
 
 ### Added
 - BroadcastSpatialJoin tests
 
 ### Changed
-- version to 1.1.3-SNAPSHOT
+- version to 1.2.1
 - Scala 2.12, Spark 2.4, JTS from locationtech v1.16
 
 ### Fixed
@@ -15,9 +32,6 @@ Fixed some bugs, added tests, build Spark 2.4 + Scala 2.12 version.
 
 ### Removed
 - limitation that don't allow to use extra condition with `NearestD` predicate
-
-### Deprecated
-- ???
 
 ## [1.1.2] - 2019-04-04
 Rewrite BroadcastSpatialJoin: added more options and flexibility, support for WGS84 (lon,lat)
@@ -47,6 +61,7 @@ Fixing bugs, Spark 2.0.2, etc.
 ## 1.0 - 2015-10-02
 Initial release
 
-[Unreleased]: https://github.com/vasnake/SpatialSpark/compare/v1.1.2...vasnake:broadcast-spatial-join
+[Unreleased]: https://github.com/vasnake/SpatialSpark/compare/v1.2.1...vasnake:broadcast-spatial-join
+[1.2.1]: https://github.com/vasnake/SpatialSpark/compare/v1.1.2...v1.2.1
 [1.1.2]: https://github.com/vasnake/SpatialSpark/compare/f9f726df75fe8e6113692b923a5cc6751112a982...v1.1.2
 [1.1.1-beta-SNAPSHOT]: https://github.com/vasnake/SpatialSpark/compare/1.0...f9f726df75fe8e6113692b923a5cc6751112a982